requirements.txt
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10.5
django-bootstrap3==7.0.1
django-extensions==1.7.4
django-filter==1.0.1
django-test-without-migrations==0.4
djangorestframework==3.5.3
Markdown==2.6.7
python-decouple==3.0
six==1.10.0
static3==0.7.0
